LayoutTests/fast/animation/request-animation-frame-iframe.html

<!DOCTYPE html>
<html>
<head>
<link rel="stylesheet" href="../js/resources/js-test-style.css">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
    <p id="description"></p>
    <div id="console"></div>
    <iframe src="script-tests/request-animation-frame-subframe.html" width="700" height="500"></iframe>
    <script type="text/javascript" charset="utf-8">
        description("Tests requestAnimationFrame in an iframe");

        var callbackInvoked = false;
    
        if (window.layoutTestController) {
            layoutTestController.dumpAsText();
            layoutTestController.waitUntilDone();
        }

        // Called from subframe.
        function doDisplay()
        {
            if (window.layoutTestController)
                layoutTestController.display();
        }
    
        function doCheckResult()
        {
            shouldBeTrue("callbackInvoked");
        }
    
        function doTestDone()
        {
            isSuccessfullyParsed();
            if (window.layoutTestController)
                layoutTestController.notifyDone();
        }

        var successfullyParsed = true;
    </script>

    <script src="../js/resources/js-test-post-function.js"></script>
</body>
</html>
