third_party/blink/web_tests/fast/animation/request-animation-frame-iframe-expected.txt

Tests requestAnimationFrame in an iframe

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS callbackInvoked is true
PASS successfullyParsed is true

TEST COMPLETE
